evaluation/requirements.txt
gdown

numpy

pillow

torch

torchnet

torchvision
